+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Support passing Node resources in JS functions
 - Add chaining methods to the `JsFunction` class
+- Add an `async()` method to the `JsFunction` class to allow developers to write `await` instructions in their JS code
 
 ### Deprecated
 - Deprecate the `JsFunction::create` method in favor of the new chaining methods